订单列表， bug修复
1、订单列表到订单详情，添加进一步的网络请求操作，确保订单数据显示的正确性
2、登录操作的bug修复

+++ EBus/Classes/Transfer/Controller/EBTransferController.m
@@ -195,7 +195,7 @@
 }
 #pragma mark - EBTransferLineCellDelegate
 - (void)transferLineOutTicktet:(EBTransferLineCell *)transeferLine transerModel:(EBTransferModel *)model type:(EBTicketType)type{
-    EBLog(@"%@, %ld",NSStringFromSelector(_cmd), type);
+    EBLog(@"%@, %ld",NSStringFromSelector(_cmd), (unsigned long)type);
     if (type == EBTicketTypeOfOut) {
         self.lineMapView.hidden = YES;
         self.tipView.hidden = NO;
